fix(runtime): redact kimi tool events
